emails: add central `prepareEmail` function
