+++ src/db/emails.js
@@ -1,3 +1,62 @@
+const crypto = require("crypto");
+
+const mail = require("@sendgrid/mail");
+
+function makeMailService() {
+  if (process.env.SENDGRID_TOKEN == null) {
+    throw new Error("missing SENDGRID_TOKEN environment variable");
+  }
+  const ms = new mail.MailService();
+  ms.setApiKey(process.env.SENDGRID_TOKEN);
+  return ms;
+}
+
+// Records an email delivery in the `email_log` table, and returns a callback
+// that can be invoked with `.send()` to actually send the email with SendGrid.
+// The caller can send the email immediately or first wait for the transaction
+// to commit, depending on which fault tolerance semantics are most
+// appropriate. The callback returns a `Promise<void>`.
+async function prepareEmail({
+  client,
+  topic,
+  email,
+  templateId,
+  templateData,
+}) {
+  await client.query(
+    `
+    INSERT INTO email_log (message_id, create_time, topic, to_email, template_id, template_data)
+    VALUES ($1::uuid, now(), $2::text, $3::text, $4::text, $5::jsonb)
+    `,
+    [
+      crypto.randomBytes(16),
+      topic,
+      email,
+      templateId,
+      JSON.stringify(templateData),
+    ]
+  );
+  const send = async () => {
+    if (process.env.NODE_ENV === "test") {
+      throw new Error(`refusing to send email in test (topic ${topic})`);
+    }
+    await makeMailService().send({
+      from: {
+        email: "noreply@archipelago.art",
+        name: "Archipelago",
+      },
+      templateId,
+      personalizations: [
+        {
+          to: [{ email }],
+          dynamicTemplateData: templateData,
+        },
+      ],
+    });
+  };
+  return { send };
+}
+
 // Adds a new email address to the signups list. Returns `true` if this made a
 // change or `false` if the email already existed in the database. Idempotent.
 async function addEmailSignup({ client, email }) {
@@ -24,6 +83,7 @@ async function getEmailSignups({ client }) {
 }
 
 module.exports = {
+  prepareEmail,
   addEmailSignup,
   getEmailSignups,
 };
